feat: open logs on shacking motion + code review notes
